+++ src/main/java/hu/ibello/requirements/FunctionalitiesCoverage.java
@@ -15,13 +15,14 @@
  */
 package hu.ibello.requirements;
 
+import java.util.regex.Pattern;
+
 /**
- * Contains the coverage of multiple requirements.
+ * Contains the coverage of multiple functionalities.
  * 
- * @author Kornél Simon
- * @see RequirementHandler#calculateTestCoverage(String)
- * @see RequirementHandler#calculateExampleCoverage(String)
+ * @see FunctionalityHandler#calculateTestCoverage(Pattern)
+ * @see FunctionalityHandler#calculateExampleCoverage(Pattern)
  */
-public class RequirementsCoverage extends ElementsCoverage<RequirementCoverage> {
+public class FunctionalitiesCoverage extends ElementsCoverage<FunctionalityCoverage> {
 
 }