+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## 1.24.0
 
 - Minimum required java version: 11
+- New `Browser.javascript()` method to run JavaScipt code
+- The `TestRun` class has a new `threads` property.
+- New `FunctionalityHandler.calculateTestCoverage(String)` method to calculate test coverage of functionalities
+- New `FunctionalityHandler.calculateExampleCoverage(String)` method to calculate example coverage of functionalities
+- New `Functionality.getLevel()` method
+- New `IbelloReportReader` interface for reading non-ibello format test results
+- New `TestResultLoader.loadTestResultFromDirectory(...)` and `TestResultLoader.loadTestResultFromFile(File)` methods
 
 ## 1.23.0